\name{bone}
\docType{data}
\alias{bone}
\title{Bone mineral content data}
\description{
The mineral density of three dominant and nondominant of bones measured in \eqn{g/cm^2} \cite{johnson1999}.}
\usage{data(bone)}
\format{A text file with 6 columns.}
\encoding{latin1}
\references{
R. A. ArnoldJohnson and D. W. Wichern 1999. \emph{Applied Multivariate Analysis}, Prentice-Hall, New Jersey.
}
\keyword{datasets}
\examples{
data(bone)
}


