\name{fatigue}
\docType{data}
\alias{fatigue}
\title{Fatigue data}
\description{
A set of 101 observations obtained by Birnbaum and Saunders(1969) from fatigue life of 6061-T6 aluminum coupons cut parallel to the direction of rolling and oscillated at 18 cycles per second (cps).}
\usage{data(fatigue)}
\format{A text file with 1 column.}
\encoding{latin1}
\references{
Z. W. Birnbaum and S. C. Saunders 1969. Estimation for a family of life distributions with applications to fatigue. \emph{Journal of Applied Probability}, 328-347.
}
\keyword{datasets}
\examples{
data(fatigue)
}
