\name{plasma}
\docType{data}
\alias{plasma}
\title{Plasma survival data}
\description{
The plasma survival data contains the Survival times of plasma cell myeloma for 112 patients, see Carbone et al. (1967).}
\usage{data(plasma)}
\format{A text file with 4 columns.}
\encoding{latin1}
\references{
P. P. Carbone, L. E. Kellerhouse, and E. A. Gehan 1967. Plasmacytic myeloma: A study of the relationship of survival to various clinical manifestations and anomalous protein type in 112 patients. \emph{The American Journal of Medicine}, 42 (6), 937-48.}
\keyword{datasets}
\examples{
data(plasma)
}
